Time: 3 ms (97.79%), Space: 40.2 MB (65.09%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.java
@@ -5,7 +5,7 @@ class Solution {
         int r = (int)Math.sqrt(c);
         
         while(l <= r){
-            long sum = l*l + r*r;
+            int sum = l*l + r*r;
             if(sum == c){
                 return true;
             }
